- SuperBitmap::drawHorizLine() no longer tries to DMA_Force() if the width of the line is 1.   - Fixed size of data stored in SuperBitmap floodFill's stack.

+++ Woopsi/woopsi/superbitmap.h
@@ -224,7 +224,7 @@ protected:
 	 * @param y The y co-ordinate of the pixel from the stack (output by the function).
 	 * @param stack Pointer to the stack to pop from.
 	 */
-	bool popStack(s16* x, s16* y, DynamicArray<u16>* stack);
+	bool popStack(s16* x, s16* y, DynamicArray<s32>* stack);
 	
 	/**
 	 * Push new co-ordinates onto the stack.
@@ -232,7 +232,7 @@ protected:
 	 * @param y The y co-ordinate of the pixel to add to the stack.
 	 * @param stack The stack to add a new value to.
 	 */
-	void pushStack(s16 x, s16 y, DynamicArray<u16>* stack) ;
+	void pushStack(s16 x, s16 y, DynamicArray<s32>* stack) ;
 	
 	/**
 	 * Clip the supplied rectangular dimensions to the size of the internal bitmap.
